requirements.txt
# Core Flask and extensions
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.8
MarkupSafe==2.1.2

# Flask extensions
flask-restx==1.3.0
flask-pymongo==3.0.1

# Database
pymongo==4.6.1
dnspython==2.7.0

# API and HTTP
requests==2.26.0
urllib3==1.26.18
charset-normalizer==2.0.0

# Scheduling
APScheduler==3.11.0
pytz==2025.2
tzlocal==5.3.1

# Environment and configuration
python-dotenv==0.19.0
